Fix review comment: Consider moving to common header in utility.swap since it is duplicated with the next file.
